requirements.txt
numpy>=1.24
matplotlib>=3.7
pyyaml>=6.0
pandas>=2.0
scipy>=1.11